// frontend/src/generated/model/uploadPurpose.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * TacoLearn API
 * Backend API template for the TacoLearn application.
 *
 * ## Response Format
 * All responses follow a standardized format:
 * - **status**: HTTP status code
 * - **message**: Human-readable message
 * - **data**: Response data (object or array)
 * - **pagination**: (Optional) Pagination metadata for list endpoints
 * OpenAPI spec version: 1.0.0
 */

/**
 * Mục đích upload. Quyết định thư mục lưu trữ, quyền truy cập, giới hạn dung lượng và định dạng file cho phép.
 */
export type UploadPurpose = typeof UploadPurpose[keyof typeof UploadPurpose];


export const UploadPurpose = {
  BUILDING_IMAGE: 'BUILDING_IMAGE',
  ROOM_IMAGE: 'ROOM_IMAGE',
  USER_AVATAR: 'USER_AVATAR',
  USER_ID_CARD_FRONT_PHOTO: 'USER_ID_CARD_FRONT_PHOTO',
  USER_ID_CARD_BACK_PHOTO: 'USER_ID_CARD_BACK_PHOTO',
  USER_PORTRAIT_PHOTO: 'USER_PORTRAIT_PHOTO',
  PAYMENT_RECEIPT_IMAGE: 'PAYMENT_RECEIPT_IMAGE',
} as const;